feat: добавление формы для генерации сообщений коммитов с поддержкой AI
- Добавлена новая форма `CommitMessageForm` для создания коммитов с поддержкой генерации сообщений с помощью AI.
- Обновлен `package.json` для изменения системного промта AI.
- Удалена старая форма `CommitForm` и заменена на новую.
- Изменен метод `getDiff` в `GitManager` для получения различий между изменениями.
